Cut down optics-th to use a subset of Language.Haskell.TH.Optics
