+++ optics-th/src/Optics/TH/Internal/Sum.hs
@@ -16,7 +16,7 @@ import qualified Data.Set as S
 import qualified Language.Haskell.TH.Datatype as D
 
 import Data.Set.Optics
-import Language.Haskell.TH.Optics
+import Language.Haskell.TH.Optics.Internal
 import Optics.Core hiding (cons)
 import Optics.TH.Internal.Utils
